[Safety Commit] added  part of error handling for remove method

+++ src/users/users.service.ts
@@ -36,12 +36,18 @@ export class UsersService {
   }
 
   public async update(id: string, updateUserDto: UpdateUserDto) {
-    // because sqlite3 does not return a number of affected rows I have to check if the entry also exists
+    // because sqlite3 does not return a number of affected rows I have to check if the entry also exists https://stackoverflow.com/questions/24030383/getting-the-rows-affected-by-update-in-sqlite3-without-extra-query
     await this.usersRepository.update(id, updateUserDto);
     return this.usersRepository.findOne(id);
   }
 
-  public async remove(id: number) {
+  public async remove(id: string) {
+    // because sqlite3 does not return a number of affected rows I have to check if the entry also exists https://stackoverflow.com/questions/24030383/getting-the-rows-affected-by-update-in-sqlite3-without-extra-query
+    const existingUser = await this.usersRepository.findOne(id);
+    if (!existingUser) {
+      return false;
+    }
     await this.usersRepository.delete(id);
+    return true;
   }
 }
